pyproject.toml
[project]
name = "tradingagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "akshare>=1.16.98",
    "backtrader>=1.9.78.123",
    "chainlit>=2.5.5",
    "chromadb>=1.0.12",
    "eodhd>=1.0.32",
    "feedparser>=6.0.11",
    "finnhub-python>=2.4.23",
    "langchain-anthropic>=0.3.15",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.1.0,<2.3.0",
    "parsel>=1.10.0",
    "plotly>=6.3.0",
    "praw>=7.8.1",
    "pytz>=2025.2",
    "questionary>=2.1.0",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "stockstats>=0.6.5",
    "streamlit>=1.49.1",
    "tqdm>=4.67.1",
    "tushare>=1.4.21",
    "typing-extensions>=4.14.0",
    "yfinance>=0.2.63",
    "darts[torch]>=0.29.0",
    "neuralforecast>=1.7.0",
    "openbb>=4.1.0",
    "prophet>=1.1.4",
    "ta-lib>=0.4.28",
    "python-docx>=1.1.0",
    "arch>=7.2.0",
    "statsmodels>=0.14.5",
    "scipy>=1.15.3",
    "cvxpy>=1.7.3",
    "pmdarima>=2.0.4",
    "ecos>=2.0.14",
    "gradio>=5.0.0",
    "ibm-watsonx-ai>=1.1.0",
]
